Initial implementation for GET and POST methods
The API offers two ways of doing the requests - either by passing all
the data and getting back QNetworkReply* and handling the responses and
errors manually, or passing all the data together with a completion
lambda, which will be passed two arguments - a QJsonDocument with the
JSON server reply and a int statusCode that will contain the HTTP code
of the response.

The second method is particularly useful for writing compact code with
handling the response in-place through lambdas, for example:

```
m_myServer->getUser([](const QJsonDocument &json, int statusCode) {
    // handle your json response
});

// The getUser() implementation in MyServer which extends ServerBase
MyServer::getUser(OnCompleted slot) {
    get("/user/me/", Parameters(), Parameters(), slot);
}
```

This will trigger the lambda (`OnCompleted slot`) once the GET request
finishes and the server reply is ready.