Allow the QObject to specify what delimiters the JSON is using
This enables the JSON with underscored properties (eg. created_at) to be
mapped to camelCase properties of the C++ class and vice versa.

A QObject with `Q_PROPERTY(int jsonDelimiter READ jsonDelimiter)` should
return in the getter which delimiter the JSON is using and the
QObjectHelper de/serializer will convert accordingly.