Let QObjectHelper also process child objects recursively
This now allows for full object hierarchy to be converted to and from
JSON. If the property for the subobject is a QObject-derived-pointer and
the QVariant value is a map (ie. a JSON subobject), QObjectHelper will
process that subobject recursively.

The downside is that this requires the metaobjects to be explicitely
registered before being first used (using `qRegisterMetaType<>()`) but
it's a small tradeoff.